Applied patch #137 by Christopher Phillips: Fixed 31 warnings in the OS X Cocoa UI code. Fixed bug where CPU window visibility couldn't be toggled by the menu item unless it had already been loaded. Fixed bug where CPU window contents would remaing blank unless it was already visible at startup. (issues #243 and 252?)
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@32214 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/unix/macosx/cocoa/dialog/memorywindowcontroller.m
@@ -238,7 +238,7 @@
     return theValue;
 }
 
-- (int)numberOfRowsInTableView:(NSTableView *)aTableView
+- (NSInteger)numberOfRowsInTableView:(NSTableView *)aTableView
 {
     if(data == nil) {
         return 0;
